refactor: split _markitdown.py into modular components
