+++ src/markitdown/converters/__init__.py
@@ -1,0 +1,3 @@
+from . import archive, base, document, media, text, web
+
+__all__ = ["archive", "base", "document", "media", "text", "web"]
