+++ src/markitdown/exceptions.py
@@ -1,0 +1,6 @@
+class FileConversionException(BaseException):
+    pass
+
+
+class UnsupportedFormatException(BaseException):
+    pass
